+++ functions/GetStaffList.fnc.php
@@ -66,7 +66,7 @@ function GetStaffList( &$extra = [] )
 			FROM staff_fields cf,staff_field_categories sfc
 			WHERE ((SELECT VALUE
 				FROM program_user_config
-				WHERE TITLE=cast(cf.ID AS VARCHAR(10))
+				WHERE TITLE=cast(cf.ID AS char(10))
 				AND PROGRAM='StaffFieldsView'
 				AND USER_ID='" . User('STAFF_ID') . "')='Y'" .
 				( ! empty( $extra['staff_fields']['view'] ) ?