+++ functions/Search.fnc.php
@@ -315,7 +315,7 @@ function Search( $type, $extra = null )
 				AND NOT EXISTS(SELECT ''
 					FROM program_user_config
 					WHERE PROGRAM='StaffFieldsSearch'
-					AND TITLE=cast(cf.ID AS VARCHAR(10))
+					AND TITLE=cast(cf.ID AS char(10))
 					AND USER_ID='" . User( 'STAFF_ID' ) . "'
 					AND VALUE='Y')
 				AND cf.TYPE<>'files'
@@ -334,7 +334,7 @@ function Search( $type, $extra = null )
 					LIMIT 1)='Y'
 				AND (SELECT VALUE
 					FROM program_user_config
-					WHERE TITLE=cast(cf.ID AS VARCHAR(10))
+					WHERE TITLE=cast(cf.ID AS char(10))
 					AND PROGRAM='StaffFieldsSearch'
 					AND USER_ID='" . User( 'STAFF_ID' ) . "'
 					LIMIT 1)='Y'
@@ -355,7 +355,7 @@ function Search( $type, $extra = null )
 				AND NOT exists(SELECT ''
 					FROM program_user_config
 					WHERE PROGRAM='StudentFieldsSearch'
-					AND TITLE=cast(cf.ID AS VARCHAR(10))
+					AND TITLE=cast(cf.ID AS char(10))
 					AND USER_ID='" . User( 'STAFF_ID' ) . "'
 					AND VALUE='Y')
 				AND cf.TYPE<>'files'
@@ -374,7 +374,7 @@ function Search( $type, $extra = null )
 					LIMIT 1)='Y'
 				AND (SELECT VALUE
 					FROM program_user_config
-					WHERE TITLE=cast(cf.ID AS VARCHAR(10))
+					WHERE TITLE=cast(cf.ID AS char(10))
 					AND PROGRAM='StudentFieldsSearch'
 					AND USER_ID='" . User( 'STAFF_ID' ) . "'
 					LIMIT 1)='Y'